Add "mood" tag to excluded tag list
The mood tag set by the AcoustBrainz plugin has values which are either e.g. "Not happy" or "Happy".

We do not want these sorted alphabetically but instead left in the same order as they are presented by AcoustBrainz.

Note: The tag name is "mood" by default, but the user can change it in the ui. If the user changes it this exclusion will not be effective.